backend/requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
python-multipart==0.0.9
python-dotenv==1.0.1

# Document parsing
pymupdf==1.25.5
python-docx==1.1.0

# Embeddings
sentence-transformers==2.7.0

# Vector store
faiss-cpu==1.9.0.post1

# LLM
groq>=0.13.0

# Utilities
numpy>=2.0.0
